requirements.txt
celery[redis]==4.4.2
Django==3.2.5
djangorestframework>=3.11.1
djangorestframework-jwt==1.11.0
django-bootstrap4>=2.2.0
django-ckeditor==5.9.0
django-compressor==2.4
django-compressor-toolkit==0.6.0
django-redis>=4.12.1
dnspython==1.16.0
gunicorn==20.0.4
Pillow==9.0.1
